requirements.txt
streamlit
pandas
plotly-express
numpy
matplotlib
seaborn
scikit-learn
joblib
jsonschema